Add files via upload
I started with the help of scimba.com, where I only had the Google image, input and the 2 main buttons on my website. After that was done I have cloned it myself. The CSS is a mess but it was a great manner for me to learn working with styling of the <divs>.